seperated menus, keys, windowrules setup
